change page type to int - issue #23

+++ src/Gridify/GridifyQuery.cs
@@ -5,7 +5,7 @@ namespace Gridify
       public GridifyQuery()
       {
       }
-      public GridifyQuery(short page, int pageSize, string filter, string? orderBy = null)
+      public GridifyQuery(int page, int pageSize, string filter, string? orderBy = null)
       {
          Page = page;
          PageSize = pageSize;
@@ -13,7 +13,7 @@ namespace Gridify
          Filter = filter;
       }
 
-      public short Page { get; set; }
+      public int Page { get; set; }
       public int PageSize { get; set; }
       public string? OrderBy { get; set; }
       public string? Filter { get; set; }
